Making it possible to use another JSON library than `jsx`
We already use `jiffy` in our project to handle JSONs, which makes
integrating `erlastic_search` a pain, since it uses `jsx`, which
has a different way of representing JSONs in Erlang, for example:

```
1> SimpleJson = <<"{\"key\":\"value\"}">>.
<<"{\"key\":\"value\"}">>
2> jiffy:decode(SimpleJson).
{[{<<"key">>,<<"value">>}]}
3> jsx:decode(SimpleJson).
[{<<"key">>,<<"value">>}]
```

This patch makes it possible to use any JSON library, as long as it defines
the two following callbacks:

```erlang
-callback encode(json()) -> binary().
-callback decode(binary()) -> json().
```

This setting is done at compile time, by defining an `ERLASTIC_SEARCH_JSON_MODULE`
environment variable containing the name of the module containing the two
callbacks above.

Updated the specs to account for this change.