Introduce new search function 'multi_search/2'
'multi_search/2' allows to execute multiple search requests using
the Elasticsearch endpoint '_msearch'. The function interface is
inspired by the 'bulk_index_docs' function with the difference that
the index and type parameters are missing. If needed these parameters
must be defined through the headers, as these are optionally
according to the Elasticsearch documentation.